test(raft): move tests that verifies network partition
